Adding task complete command and admin check on delete

+++ WebAPI/Controllers/TagController.cs
@@ -1,14 +1,17 @@
 ﻿using Application;
 using MediatR;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace WebAPI.Controllers
 {
+    [Authorize]
     [ApiController]
     [Route("[controller]")]
     public class TagController : ControllerBase
@@ -42,8 +45,12 @@ namespace WebAPI.Controllers
 
         [HttpDelete]
         [Route("Delete")]
-        public async Task<IActionResult> DeleteTag([FromRoute]string id)
+        public async Task<IActionResult> DeleteTag(string id)
         {
+            var userID = User.Claims.Where(a => a.Type == ClaimTypes.Email).FirstOrDefault().Value;
+            if (userID != "admin@maus.ba") //TODO get from db.
+                return Unauthorized();
+
             var query = new DeleteTagCommand(id);
             var result = await _mediator.Send(query);
             return Ok(result);
